Connected Standby Mode
Allows the user to delay the activation of Airplane mode in standby.
The default delay is one second, can be configured by sending an
intent to BonovoHandle with the delay amount in milliseconds.

This allows the use of Wifi and Bluetooth in sleep until the delay
times out after which airplane mode will activate.