Bluetooth
This patch fixes some issues created with the previous commit and improves reliablility of bluetooth music playback.

* Fixes playback errors when the Android box wakes from sleep
* Adds an image to the track information (visible in apps like Jacks Music Widget)
* tracks when the HFP and A2DP profiles are connected and refuses a command if the appropriate profile is not conncted.
* Detects playback errors and automatically resets the Bluetooth module.  If this occurs when music was playing it should be restarted as soon as the module finishess it's restart.
* In a phone call allow the connected phone to send us the other persons name from it's contacts and show that instead.
* Get Audio focus sooner when a phone call starts.  This fixes issues like Siri not playing it's beeps to indicate it is ready to listen.

The Bluetooth module seems to have problems if a command is issued to it before it has finished starting the required profile.  This update should take care of the most common causes of this happening.