Reduce the update rate from every second to every two seconds. Slows down reaction time when the track changes but the BT module seems much happier and far less likely to throw errors in this case.
also stop updating the track position in the remotecontrolclient and
only update the play status when it changes.  removes a lot of excess
debugging messages.

Instead of resetting the BT module as soon as an error occurs, now we track
how often they happen and only reset it after 3 errors. No need to reset if
it is able to handle it on it's own and resume normal operation