Bluetooth Split
Includes the service changes required by the new bluetoothmusic app and removes the old music player from BluetoothSettings.
There are now separate launcher icons for Music, Phone and Settings.  Settings is still part of the main package due to being
tied to the main service still.

Also pulls in a fix to the Bluetooth init code that was too hard to spearate out on it's own.

All still a work in progress..