Add a shutdown watchdog to BonovoHandle
I'm sick of coming back to my car and finding the MCU didn't shutdown the Android system as it should have.

This is an attempt to fix that by adding a timer to BonovoHandle.  It is scheduled 1 minute after the MCU should have powered the box down.  If it's still running at that time, call the ANdroid shutdown intent. Since BonovoHandle is a system app this should work.
Only tested on the workbench so far, not in the car yet.

Also ging to clean up some of the cruft in this file soon.  There is a heap of code we don't use and the formatting is all over the place.  I don't expect to do another merge from Newsmy anytime soon so i think it's time we start making this a bit nicer place to work on.