Create a concept of capabilities and presets
This allows individual features (Capability) of the devices to be toggleable,
and a set of pre-defined presets with capabilities enabled/disabled. Those
presets coincide with the NU- and NR-series of devices.

The user is able to choose which preset they want to enable, or optionally
take full control over the capabilities to enable/disable specific features
that he wants.

Eventually this concept should probably move into the com.bonovo framework,
so that other apps and system services can make use of the presets and
capabilities.

For now, nothing is actually using the capabilities, but coming soon, we
should update the MCU Settings to enable/disable certain features in the UI
based on the capability selections.

Issue: Nu3001/xdAuto#8