Add a SCAN feature
On long-press of the STEP button, it will initiate an automated SCAN mode,
where it will STEP right to the next frequency, and then pause for N seconds.
After N seconds, it will continue to the next frequency, until stopped (or
until it makes it all the way back around to the starting point).