Autplay last media app
This still needs work, but it captures the last playing media app at shutdown.
Assuming the media package is recognized by our media intent system.
The code is capturing at least Radio & Google Play Music correctly.
It also starts the app on boot, but the autoplay part isn't working.
It seems to be a timing issue on the boot/pre-launcher receiver.

Signed-off-by: Mike Wilson <mwilson@acesappraisal.com>